+++ CHANGELOG.md
@@ -1,34 +1,50 @@
+## 1.5.0 (2018-08-10)
+
+Features:
+   - Add Single-App-to-Multi-App-Platform project.
+   - Add SNI-LED project.
+   - Add Multi-App-Security-Manager project.
+
 ## 1.4.0 (2018-03-01)
+
 Features:
    - Add FoundationLibrary-Get-Started project.
 
 ## 1.3.1 (2018-01-31)
+
 Bug fixes:
    - Update Proguard-Get-Started.
    - Update Storage-Get-Started.
 
 ## 1.3.0 (2017-09-04)
+
 Features:
    - Add Proguard-Get-Started project.
 
 ## 1.2.1 (2017-08-17)
+
 Features:
    - Update Storage-Get-Started project.
 
 ## 1.2.0 (2017-08-07)
+
 Features:
   - Add Widgets-Get-Started project.
   - Add Storage-Get-Started project.
 
 ## 1.1.0 (2016-08-25)
+
 Features:
   - Add MicroUI-Get-Started project.
-  
+
 ## 1.0.0 (2016-07-01)
+
 Features:
   - Initial revision.
 
-[//]: # (Markdown)
-[//]: # (Copyright 2017-2018 IS2T. All rights reserved.)
-[//]: # (For demonstration purpose only.)
-[//]: # (IS2T PROPRIETARY. Use is subject to license terms.)
+<!--
+	Markdown
+	Copyright 2017-2018 IS2T. All rights reserved.
+	For demonstration purpose only.
+	IS2T PROPRIETARY. Use is subject to license terms.
+-->